aboutページの画面遷移を追加
