--- src/app/about.html
@@ -1,7 +1,0 @@
-<div class="content">
-  <p class="center" style="text-align:center;font-weight:bold;margin-bottom:3rem;">このアプリについて</p>
-
-  <div class="center" style="text-align:center;margin-bottom:2rem;">アプリバージョン<br>xxxxxxxx</div>
-
-  <div class="center"  style="text-align:center;margin-bottom:2rem;">バージョンノート<br>xxxxxxxx</div>
-</div>
