--- src/app/about.ts
@@ -1,11 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'ons-page[about]',
-  template: require('./about.html')
-})
-export class About {
-  constructor() {}
-}
-
-
