+++ src/app/app.html
@@ -1,16 +1,3 @@
 <ons-page>
-  <ons-toolbar [attr.modifier]="modifier">
-    <div class="center">お薬手帳</div>
-  </ons-toolbar>
-
-  <div class="content">
-    <ons-tabbar [attr.animation]="animation" position="auto">
-      <div class="tabbar__content"></div>
-      <div class="tabbar">
-        <ons-tab icon="fa-home" label="ホーム" [page]="home" active></ons-tab>
-        <ons-tab icon="fa-wrench" label="設定" [page]="setting"></ons-tab>
-        <ons-tab icon="fa-plus" label="追加" (click)="onPlusButtonClick()" [page]="add"></ons-tab>
-      </div>
-    </ons-tabbar>
-  </div>
+  <ons-navigator swipable [page]="page"></ons-navigator>
 </ons-page>
