--- src/app/setting.ts
@@ -1,12 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'ons-page[setting]',
-  template: require('./setting.html'),
-  styleUrls: ['/src/app/setting.css']
-})
-export class Setting {
-  constructor() {}
-}
-
-
