+++ src/app/setting/about.html
@@ -1,0 +1,11 @@
+<ons-toolbar [attr.modifier]="modifier">
+  <div class="left" (click)="onCloseClicked()"><ons-toolbar-button><ons-icon icon="fa-times"></ons-icon></ons-toolbar-button></div>
+  <div class="center">お薬手帳</div>
+</ons-toolbar>
+<div class="content">
+  <p class="center" style="text-align:center;font-weight:bold;margin-bottom:3rem;">このアプリについて</p>
+
+  <div class="center" style="text-align:center;margin-bottom:2rem;">アプリバージョン<br>xxxxxxxx</div>
+
+  <div class="center" style="text-align:center;margin-bottom:2rem;">バージョンノート<br>xxxxxxxx</div>
+</div>
