+++ src/app/setting/setting.html
@@ -5,7 +5,7 @@
         <div style="margin: 30px auto; height: 30px; background: #ffffff; text-align: center; color: #000000; border: 3px solid #000000;
 ">プライバシーポリシー</div>
         <div style="margin: 30px auto; height: 30px; background: #ffffff; text-align: center; color: #000000; border: 3px solid #000000;
-">このアプリについて</div>
+" (click)="onAboutClicked()" >このアプリについて</div>
         <div style="margin: 30px auto; height: 30px; background: #ffffff; text-align: center; color: #000000; border: 3px solid #000000;
 ">お問い合わせ</div>
     </section>
