+++ src/main.ts
@@ -10,8 +10,10 @@ import {OnsenModule} from 'ngx-onsenui';
 
 import {MyApp} from './app/app';
 import {Home} from './app/home';
-import {Setting} from './app/setting';
-import {About} from './app/about';
+import {Setting} from './app/setting/setting';
+import {SettingMain} from './app/setting/setting-main';
+import {About} from './app/setting/about';
+import {MainTab} from './app/main-tab';
 
 // Enable production mode when in production mode.
 if (process.env.NODE_ENV === 'production') {
@@ -27,12 +29,16 @@ if (process.env.NODE_ENV === 'production') {
         MyApp,
         Home,
         Setting,
+        SettingMain,
         About,
+        MainTab,
     ],
     entryComponents: [
         Home,
         Setting,
+        SettingMain,
         About,
+        MainTab,
     ],
     bootstrap: [
         MyApp,
